requirements.txt
google-generativeai
docx2python
numpy
pandas
scikit-learn
